Remove obsolete pick_folder
