+++ desktop/src/app/events/message.rs
@@ -1,0 +1,63 @@
+use iced::{widget::text_editor, Event};
+use std::path::PathBuf;
+
+use crate::app::{AppTheme, CreateTarget, HotkeyField, Language, FileEntry};
+
+#[derive(Debug, Clone)]
+pub enum Message {
+    PickFolder,
+    FolderPicked(Option<PathBuf>),
+    FilesLoaded(Vec<FileEntry>),
+    QueryChanged(String),
+    SelectFile(PathBuf),
+    FileLoaded(Result<(PathBuf, String), String>),
+    FileContentEdited(text_editor::Action),
+    SearchTermChanged(String),
+    ReplaceTermChanged(String),
+    Find,
+    ReplaceAll,
+    SaveFile,
+    FileSaved(Result<(), String>),
+    NewFileNameChanged(String),
+    NewDirectoryNameChanged(String),
+    CreateTargetChanged(CreateTarget),
+    CreateFile,
+    ConfirmCreateFile,
+    CancelCreateFile,
+    FileCreated(Result<PathBuf, String>),
+    CreateDirectory,
+    DirectoryCreated(Result<PathBuf, String>),
+    RenameFileNameChanged(String),
+    RenameFile,
+    FileRenamed(Result<PathBuf, String>),
+    RequestDeleteFile,
+    DeleteFile,
+    CancelDeleteFile,
+    FileDeleted(Result<PathBuf, String>),
+    CloseFile(usize),
+    FileClosed(Result<usize, String>),
+    RunSearch,
+    SearchFinished(Result<Vec<String>, String>),
+    RunParse,
+    ParseFinished(Result<Vec<String>, String>),
+    RunGitLog,
+    GitFinished(Result<Vec<String>, String>),
+    RunExport,
+    ExportFinished(Result<Vec<String>, String>),
+    CoreEvent(String),
+    IcedEvent(Event),
+    SaveSettings,
+    SettingsSaved,
+    OpenSettings,
+    CloseSettings,
+    ToggleDir(PathBuf),
+    ShowContextMenu(PathBuf),
+    CloseContextMenu,
+    ConfirmDiscard,
+    CancelDiscard,
+    ThemeSelected(AppTheme),
+    LanguageSelected(Language),
+    StartCaptureHotkey(HotkeyField),
+    SwitchToTextEditor,
+    SwitchToVisualEditor,
+}
