+++ desktop/src/app/events/mod.rs
@@ -1,0 +1,4 @@
+pub mod message;
+pub mod handler;
+
+pub use message::Message;
